fix!: 綜合 code review 修復 — P0 安全漏洞 + P1 正確性 + P2 品質改善
P0（安全/合規）:
- _user_locks 改用 hashed user ID 作為 key，避免記憶體保留原始 LINE user ID
- background 模式改用 settings.is_production 取代硬編碼 Render 環境變數
- 新增 LineClient.close() 與 close_line_client() 避免 HTTP 連線洩漏
- 修正 CORS allow_credentials=True + allow_origins=["*"] 違規組合

P1（正確性/健壯性）:
- LLM client 新增 per-call max_tokens 參數，預設 4096
- Gemini retry 加入 TimeoutError 處理
- RouterService LLM 呼叫加入 max_tokens 限制
- clear_all_data 補齊 PracticeSessionModel 軟刪除
- user_state_repo _get_or_create 加入 IntegrityError savepoint 重試
- 軟刪除 resurrection 時清除所有暫存欄位
- get_random_by_user 修正 exclude_ids filter 順序
- 移除 cost_service 死碼（~77 行）
- 修正 privacy.py 不一致（Claude/GPT → Claude/Gemini、加密→雜湊）
- messages.py 補齊 type hints

P2（品質/效能）:
- format_item_label 改為公開方法
- X-Request-ID 加入格式驗證（長度 ≤128、英數字 + -_）
- 移除 database.py 未使用的 get_session_dependency
- extractor prompt 跳脫 </user_input> 防止 XML tag injection
- _is_duplicate_event 改為惰性清理（僅超限時觸發）
- add_tokens 改用 UPDATE ... RETURNING 減少 3→1 次 DB round trip
- 40 處 f-string logging 改為 lazy % formatting
- cost_usd Float → Numeric(10,6)（需 migration）
- pending_delete_items Text → JSONB（需 migration）
- practice_logs 新增 item_id 單獨索引（需 migration）

測試修復:
- conftest 明確設定 APP_ENV=development 避免 background 模式干擾
- 更新 test_delete/test_router_service/test_user_profile_repo 配合改動

BREAKING CHANGE: DeleteService._format_item_label 更名為 format_item_label

+++ src/models/api_usage_log.py
@@ -8,7 +8,7 @@ from datetime import datetime
 from uuid import uuid4
 
 import sqlalchemy as sa
-from sqlalchemy import Boolean, DateTime, Float, Index, Integer, String, func
+from sqlalchemy import Boolean, DateTime, Index, Integer, Numeric, String, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -73,7 +73,7 @@ class ApiUsageLog(Base):
         comment="Output token count",
     )
     cost_usd: Mapped[float] = mapped_column(
-        Float,
+        Numeric(10, 6),
         nullable=False,
         default=0.0,
         comment="Calculated cost in USD",
